* g_krbrlm.c: Remove unused static variable krb_conf
	* g_svc_in_tkt.c (krb_svc_init, krb_svc_init_preauth): Declare as
	returning int.

	* gethostname.c: Include unistd.h for gethostname() prototype.

	* getst.c: Include unistd.h for read() prototype.

	* in_tkt.c (in_tkt): Cast arguments to debugging printf to int
	from uid_t to match format statement.

	* kname_parse.c: Declare k_isname() and k_isinst() as returning
	int. Cleanup assigments in conditionals.

	* kuserok.c (kuserok): Cleanup assignment in conditional.

	* log.c (krb_set_logfile): Declare function as void.
	* klog.c (kset_logfile): Likewise.

	* pkt_clen.c (pkt_clen): Declare as returning int.
	* kntoln.c (krb_kntoln): Likewise.
	* fgetst.c (fgetst): Likewise.

	* rd_req.c: Declare local variable only if KRB_CRYPT_DEBUG defined.

	* recvauth.c: Include stdlib.h and unistd.h for read() and atoi()
	prototypes.

	* send_to_kdc.c: Include unistd.h for close() prototype.

	* sendauth.c (krb_sendauth): Clean up assignment in conditional.

	* tkt_string.c (tkt_string): Likewise

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13256 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb4/in_tkt.c
@@ -156,7 +156,7 @@ in_tkt(pname,pinst)
 	    return(KFAILURE);
 	} else
 	    if (krb_debug)
-		printf("swapped UID's %d and %d\n",metoo,me);
+		printf("swapped UID's %d and %d\n",(int) metoo, (int) me);
     }
     /* Set umask to ensure that we have write access on the created
        ticket file.  */
@@ -171,7 +171,7 @@ in_tkt(pname,pinst)
 	    return(KFAILURE);
 	} else
 	    if (krb_debug)
-		printf("swapped UID's %d and %d\n",me,metoo);
+		printf("swapped UID's %d and %d\n", (int) me, (int) metoo);
     }
     if (tktfile < 0) {
 	if (krb_debug)
